+++ pioneer_modules/src/drive.cpp
@@ -223,8 +223,9 @@ rcl_interfaces::msg::SetParametersResult Drive::dynamicParametersCallback(
       }
     } else if (type == ParameterType::PARAMETER_INTEGER) {
       if (name == plugin_name_ + ".velocity_timeout") {
-        vel_timeout_ = rclcpp::Duration::from_seconds(parameter.as_int() / 1000.0);
-        RCLCPP_INFO(logger_, "The parameter velocity_timeout is set to: [%i]", parameter.as_int());
+        int timeout = parameter.as_int();
+        vel_timeout_ = rclcpp::Duration::from_seconds(timeout / 1000.0);
+        RCLCPP_INFO(logger_, "The parameter velocity_timeout is set to: [%i]", timeout);
       }
     }
   }